refactor: [4단계] 상위 패키지에 domain 추가
