--- src/main/java/roomescape/constant/AuthConstant.java
@@ -1,19 +1,0 @@
-package roomescape.constant;
-
-public enum AuthConstant {
-
-    TOKEN("token"),
-    ROLE("role"),
-    NAME("name"),
-    ADMIN("admin");
-
-    private String value;
-
-    AuthConstant(final String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
