+++ src/main/java/roomescape/domain/member/MemberController.java
@@ -1,0 +1,26 @@
+package roomescape.domain.member;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import roomescape.domain.member.dto.MemberRequest;
+import roomescape.domain.member.dto.MemberResponse;
+
+@RestController
+public class MemberController {
+
+    private MemberService memberService;
+
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @PostMapping("/members")
+    public ResponseEntity createMember(@RequestBody MemberRequest memberRequest) {
+        MemberResponse member = memberService.createMember(memberRequest);
+        return ResponseEntity.created(URI.create("/members/" + member.getId())).body(member);
+    }
+}
