+++ src/main/java/roomescape/domain/member/dto/LoginMember.java
@@ -1,0 +1,8 @@
+package roomescape.domain.member.dto;
+
+public record LoginMember(
+    Long id,
+    String name,
+    String role
+) {
+}
