+++ src/main/java/roomescape/domain/member/dto/MemberLoginCheckResponse.java
@@ -1,0 +1,6 @@
+package roomescape.domain.member.dto;
+
+public record MemberLoginCheckResponse(
+        String name
+) {
+}
