+++ src/main/java/roomescape/domain/member/dto/MemberLoginRequest.java
@@ -1,0 +1,7 @@
+package roomescape.domain.member.dto;
+
+public record MemberLoginRequest(
+        String password,
+        String email
+) {
+}
