+++ src/main/java/roomescape/domain/member/dto/MemberRequest.java
@@ -1,0 +1,19 @@
+package roomescape.domain.member.dto;
+
+public class MemberRequest {
+    private String name;
+    private String email;
+    private String password;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
