+++ src/main/java/roomescape/domain/member/dto/MemberResponse.java
@@ -1,0 +1,25 @@
+package roomescape.domain.member.dto;
+
+public class MemberResponse {
+    private Long id;
+    private String name;
+    private String email;
+
+    public MemberResponse(Long id, String name, String email) {
+        this.id = id;
+        this.name = name;
+        this.email = email;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+}
