+++ src/main/java/roomescape/domain/theme/Theme.java
@@ -1,0 +1,33 @@
+package roomescape.domain.theme;
+
+public class Theme {
+    private Long id;
+    private String name;
+    private String description;
+
+    public Theme() {
+    }
+
+    public Theme(Long id, String name, String description) {
+        this.id = id;
+        this.name = name;
+        this.description = description;
+    }
+
+    public Theme(String name, String description) {
+        this.name = name;
+        this.description = description;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
