+++ src/main/java/roomescape/domain/time/AvailableTime.java
@@ -1,0 +1,25 @@
+package roomescape.domain.time;
+
+public class AvailableTime {
+    private Long timeId;
+    private String time;
+    private boolean booked;
+
+    public AvailableTime(Long timeId, String time, boolean booked) {
+        this.timeId = timeId;
+        this.time = time;
+        this.booked = booked;
+    }
+
+    public Long getTimeId() {
+        return timeId;
+    }
+
+    public String getTime() {
+        return time;
+    }
+
+    public boolean isBooked() {
+        return booked;
+    }
+}
