+++ src/main/java/roomescape/domain/time/Time.java
@@ -1,0 +1,27 @@
+package roomescape.domain.time;
+
+public class Time {
+    private Long id;
+    private String value;
+
+    public Time(Long id, String value) {
+        this.id = id;
+        this.value = value;
+    }
+
+    public Time(String value) {
+        this.value = value;
+    }
+
+    public Time() {
+
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
