+++ src/main/java/roomescape/global/constant/AuthConstant.java
@@ -1,0 +1,19 @@
+package roomescape.global.constant;
+
+public enum AuthConstant {
+
+    TOKEN("token"),
+    ROLE("role"),
+    NAME("name"),
+    ADMIN("admin");
+
+    private String value;
+
+    AuthConstant(final String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
