--- src/main/java/roomescape/member/dto/LoginMember.java
@@ -1,8 +1,0 @@
-package roomescape.member.dto;
-
-public record LoginMember(
-    Long id,
-    String name,
-    String role
-) {
-}
