--- src/main/java/roomescape/member/dto/MemberLoginCheckResponse.java
@@ -1,6 +1,0 @@
-package roomescape.member.dto;
-
-public record MemberLoginCheckResponse(
-        String name
-) {
-}
