--- src/main/java/roomescape/member/dto/MemberLoginRequest.java
@@ -1,7 +1,0 @@
-package roomescape.member.dto;
-
-public record MemberLoginRequest(
-        String password,
-        String email
-) {
-}
