--- src/main/java/roomescape/reservation/ReservationController.java
@@ -1,49 +1,0 @@
-package roomescape.reservation;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-import roomescape.member.dto.LoginMember;
-import roomescape.reservation.dto.ReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
-
-@RestController
-public class ReservationController {
-
-    private final ReservationService reservationService;
-
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
-    }
-
-    @GetMapping("/reservations")
-    public List<ReservationResponse> list() {
-        return reservationService.findAll();
-    }
-
-    @PostMapping("/reservations")
-    public ResponseEntity create(@RequestBody ReservationRequest reservationRequest,
-                                 LoginMember loginMember) {
-        if (reservationRequest.getDate() == null
-                || reservationRequest.getTheme() == null
-                || reservationRequest.getTime() == null) {
-            return ResponseEntity.badRequest().build();
-        }
-        ReservationResponse reservation = reservationService.save(reservationRequest, loginMember);
-
-        return ResponseEntity.created(URI.create("/reservations/" + reservation.getId())).body(reservation);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity delete(@PathVariable Long id) {
-        reservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
-}
