--- src/main/java/roomescape/time/Time.java
@@ -1,27 +1,0 @@
-package roomescape.time;
-
-public class Time {
-    private Long id;
-    private String value;
-
-    public Time(Long id, String value) {
-        this.id = id;
-        this.value = value;
-    }
-
-    public Time(String value) {
-        this.value = value;
-    }
-
-    public Time() {
-
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
