+++ src/test/java/roomescape/MissionStepTest.java
@@ -5,15 +5,12 @@ import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
 import java.util.HashMap;
 import java.util.Map;
-import roomescape.auth.JwtTokenManager;
-import roomescape.member.Member;
-import roomescape.reservation.dto.ReservationResponse;
+import roomescape.domain.reservation.dto.ReservationResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
